Signed-off-by: Jakayus <jakayus123@gmail.com>
Created button to be filled in later. This also fixes a bug issue.

Bug appears to occur when listview has no more items and user performs on delete. Without another item, it will crash the program. Unsure if this is a known SwiftUI issue, but appears to be the case. Otherwise more code inspection is needed.